refactor : 변수명 변경 및 setter 제거
